+++ include/game_of_life.hpp
@@ -24,7 +24,7 @@ class GameOfLife
 public:
   GameOfLife(); // Constructs a 100x100 GameOfLife object.
   GameOfLife(int rows, int cols); // Constructs a rows x cols GameOfLife object.
-  void Move(double elapsed_time); // Move forward in time.
+  void Move(float elapsed_time); // Move forward in time.
   void AddIndividual(std::unique_ptr<Individual>);
   void SetTileType(int i, int j, int type);
   Tile* GetTile(int, int);
@@ -36,16 +36,16 @@ public:
 private:
   int num_rows_{100};
   int num_cols_{100};
-  double dx_{2.0/100.0};
-  double dy_{2.0/100.0};
-  double dt_{1.0/60.0};
-  double time_accumulator_{0.0};
-  double time_factor_{1.0};
+  float dx_{2.0f/100.0f};
+  float dy_{2.0f/100.0f};
+  float dt_{1.0f/60.0f};
+  float time_accumulator_{0.0f};
+  float time_factor_{1.0f};
   std::vector<std::unique_ptr<Individual>> individuals_;
   std::vector<std::vector<Tile>> tiles_;
 
   void Initialize();
-  void UpdateTiles(double dt);
+  void UpdateTiles(float dt);
   void PositionToTile(glm::dvec2, int*, int*) const;
 
   friend Renderer;
